pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "1kh"
version = "0.1.0"
description = "ThousandHand - Autonomous business-building system"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Paul", email = "apollo.d.paradise@gmail.com" }
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "anthropic>=0.18.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
temporal = [
    "temporalio>=1.4.0",
]

[project.scripts]
1kh = "cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["cli", "core", "temporal"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
